+++ src/powershell/modules/Media/Videoscreenshot/Private/New-RunContext.ps1
@@ -52,5 +52,6 @@ function New-VideoRunContext {
         SaveFolder      = $SaveFolder
         RunGuid         = $RunGuid
         PidRegistryPath = $null
+        VlcLogPath      = $null
     }
 }